+++ config.h
@@ -158,6 +158,8 @@ struct config {
         uint32_t selection_bg;
         uint32_t url;
 
+        uint32_t dim[8];
+
         struct {
             uint32_t fg;
             uint32_t bg;
@@ -173,6 +175,7 @@ struct config {
             bool jump_label:1;
             bool scrollback_indicator:1;
             bool url:1;
+            uint8_t dim;
         } use_custom;
     } colors;